Major cleaning-up (or rather correction) of the work-flow, dependency, and other mess in the implementation of invoking an analysis.
Just to be cautious, the change was made only to ANOVA in this commitment.

More major changes are coming in two directions: (1) the same change applies to other analysis plug-ins; and (2) further correction/cleaning-up in the design of analysis/plug-in/visual plug-in is needed.

This commitment affects 5 files: GeworkbenchRoot, Anova, AnovaUI, MicroarrayUI, PluginRegistry.

git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench-web@10275 8b29bb6e-589a-435a-9a12-3ab907be3462